Minor bugfix
Have managed to connect it to the people app, albeit a rather bad
connection.

It now correctly displays events that onyl affects specific persons.

What needs to be done: Once an event is created, it should also create
an eventparticipation which will create a relation between the user and
the event