Further Improvements
A new event is always displayed, until it has been saved, even if
another view is selected

For example: if "View only note events" is clicked, and the user creates
a "meeting event", the meeting event will be created and displayed like
normal until it has been saved. Once it is saved it is no longer counted
as a new event.