Trying out custom element and started working on "balls"
Each event in the list should have a "ball" which visually shows what
type of event it is. This "ball" should be a point in the timeline
"line".

At the moment I've just added in a ball to kinda see how it should look
like. This is still only a Work In Progress